Refactor todos db handler filename
